test(agent): identity-preservation suite in its own file
PR3 review round 1, finding 4/4: the GH#2273 describe moves to
swm-recovery-identity-preservation.test.ts (swm-recovery.test.ts returns
under 1k lines and keeps the lane's transport/apply/progress rows).
Registered in the unit config; CI shards auto-discover. 23/23 green across
the pair.
