+++ packages/agent/test/swm-recovery-identity-preservation.test.ts
@@ -1,0 +1,236 @@
+/**
+ * GH#2273 part 3 — identity-preservation rows for the PRIVATE curator-recovery
+ * lane, split from `swm-recovery.test.ts` (which pins the lane's transport,
+ * apply and progress behaviors) so each file stays scannable. Same real-store
+ * methodology; the lifecycle-shaped meta-replacement helper is deliberately
+ * local to this focused suite.
+ */
+import { afterEach, describe, expect, it } from 'vitest';
+import { OxigraphStore } from '@origintrail-official/dkg-storage';
+import {
+  contextGraphWorkspaceMetaGraphUri,
+  type OperationContext,
+} from '@origintrail-official/dkg-core';
+import type { Quad } from '@origintrail-official/dkg-storage';
+import type { SyncPageResult } from '../src/sync/requester/page-fetch.js';
+import { recoverContextGraphSwm } from '../src/sync/requester/swm-recovery.js';
+import { createSharedMemorySnapshotMaterializer } from '../src/sync/requester/swm-snapshot-materializer.js';
+import type { GraphScopedSwmRecoveryDescriptor } from '../src/sync/graph-scoped-swm-recovery.js';
+import { swmFixtures } from './swm-descriptor-fixtures.js';
+
+const CG = 'ws00-recovery';
+const WS_META = contextGraphWorkspaceMetaGraphUri(CG);
+const DKG = 'http://dkg.io/ontology/';
+const ctx: OperationContext = { operationId: 'test', operationName: 'sync' } as never;
+
+function page(quads: Quad[], completed = true): SyncPageResult {
+  return { quads, bytesReceived: 0, resumedFromOffset: 0, nextOffset: quads.length, checkpointKey: 'k', completed };
+}
+
+/**
+ * GH#2273 part 3 — the private curator-recovery lane must not rotate the
+ * operation identity of a KA it SKIPPED as already materialized. Pre-fix the
+ * per-KA `continue` skipped only the graph replace; the unconditional bulk
+ * `replaceMetaForGraphAssets(graphScopedDescriptors)` then deleted the
+ * member-author's head + operation subject and installed the curator's
+ * identity for content that never changed — the single-step form of the
+ * rotation that terminally kills queued VM-publish jobs frozen on the local
+ * id. These rows wire the REAL meta-replacement implementation (the
+ * lifecycle's shape: kaUal-guarded linked-operation deletes) and the real
+ * materializer, both OPT-IN per row — the suites above deliberately run
+ * without them and pin the lane's other behaviors unchanged.
+ */
+describe('recoverContextGraphSwm preserves operation identity for skipped KAs (GH#2273)', () => {
+  const stores: OxigraphStore[] = [];
+  afterEach(async () => { await Promise.all(stores.splice(0).map((s) => s.close().catch(() => {}))); });
+
+  const swmFx = swmFixtures(CG);
+  const UAL3 = 'did:dkg:hardhat:31337/0xcccccccccccccccccccccccccccccccccccccccc/3';
+  const localShare = swmFx.share({ version: 1, operationId: 'op-local', marker: 'identity', ual: UAL3 });
+  const curatorEquivalent = swmFx.share({ version: 1, operationId: 'storage-ack-x', marker: 'identity', ual: UAL3 });
+  const curatorChanged = swmFx.share({ version: 1, operationId: 'storage-ack-y', marker: 'changed', ual: UAL3 });
+
+  /** The lifecycle implementation's shape: delete head + kaUal-owned linked operations. */
+  function realReplaceMetaForGraphAssets(store: OxigraphStore) {
+    return async (assets: readonly GraphScopedSwmRecoveryDescriptor[]) => {
+      for (const asset of assets) {
+        const linked = await store.query(
+          `SELECT DISTINCT ?op WHERE { GRAPH <${asset.metaGraph}> { ` +
+          `<${asset.headSubject}> <${DKG}shareOperationId> ?shareId . ` +
+          `?op <${DKG}shareOperationId> ?shareId ; <${DKG}kaUal> <${asset.kaUal}> . } }`,
+        );
+        const subjects = new Set<string>([asset.operationSubject]);
+        if (linked.type === 'bindings') {
+          for (const row of linked.bindings) if (row['op']) subjects.add(String(row['op']));
+        }
+        await store.deleteByPattern({ graph: asset.metaGraph, subject: asset.headSubject });
+        for (const subject of subjects) {
+          await store.deleteByPattern({ graph: asset.metaGraph, subject });
+        }
+      }
+    };
+  }
+
+  /** The production private-lane predicate: the marker-only assertionGraph ASK. */
+  function markerOnlyMaterialized(store: OxigraphStore) {
+    return async (asset: GraphScopedSwmRecoveryDescriptor) => {
+      const result = await store.query(
+        `ASK { GRAPH <${asset.metaGraph}> { <${asset.headSubject}> ` +
+        `<${DKG}assertionGraph> <${asset.assertionGraph}> . } }`,
+      );
+      return result.type === 'boolean' && result.value;
+    };
+  }
+
+  function makeIdentityBaseDeps(store: OxigraphStore, curatorMeta: Quad[]) {
+    return {
+      ctx,
+      remotePeerId: 'peer-curator',
+      contextGraphId: CG,
+      deadline: Number.MAX_SAFE_INTEGER,
+      fetchSyncPages: async (
+        _c: OperationContext, _p: string, _cg: string, _inc: boolean,
+        phase: 'data' | 'meta',
+      ): Promise<SyncPageResult> => page(phase === 'meta' ? curatorMeta : []),
+      processSharedMemoryBatch: async (dataQuads: Quad[], metaQuads: Quad[]) => ({
+        verifiedData: dataQuads,
+        verifiedMeta: metaQuads,
+        entityCreators: [],
+        droppedDataTriples: 0,
+      }),
+      store,
+      publicSnapshotStore: identitySnapshotStore(),
+      ensureContextGraph: async () => {},
+      setCheckpoint: () => {},
+      deleteCheckpoint: () => {},
+    };
+  }
+
+  /** Pre-seeded with every fixture payload so any served ref is fetchable. */
+  function identitySnapshotStore() {
+    const snapshots = new Map<string, Quad[]>();
+    for (const share of [localShare, curatorEquivalent, curatorChanged]) {
+      snapshots.set(share.digest, share.payload.map((quad) => ({ ...quad })));
+    }
+    return {
+      async putSnapshot(input: { readonly digest: string; readonly quads: readonly Quad[] }) {
+        snapshots.set(input.digest, input.quads.map((quad) => ({ ...quad })));
+        return { ref: input.digest, byteLength: 0 };
+      },
+      async getSnapshot(ref: string): Promise<Quad[] | null> {
+        return snapshots.get(ref)?.map((quad) => ({ ...quad })) ?? null;
+      },
+    };
+  }
+
+  function identityDeps(store: OxigraphStore, curatorMeta: Quad[]) {
+    return {
+      ...makeIdentityBaseDeps(store, curatorMeta),
+      replaceMetaForGraphAssets: realReplaceMetaForGraphAssets(store),
+      isGraphAssetMaterialized: markerOnlyMaterialized(store),
+      snapshotMaterializer: createSharedMemorySnapshotMaterializer({
+        store,
+        writeLocks: new Map<string, Promise<void>>(),
+        invalidateListContextGraphsCache: () => {},
+      }),
+    };
+  }
+
+  async function seedLocal(store: OxigraphStore, share = localShare): Promise<void> {
+    await store.insert(share.payload.map((quad) => ({ ...quad, graph: share.assertionGraph })));
+    await store.insert([...share.meta]);
+  }
+
+  async function headIds(store: OxigraphStore): Promise<string[]> {
+    const result = await store.query(
+      `SELECT DISTINCT ?op WHERE { GRAPH <${WS_META}> { <${localShare.headSubject}> <${DKG}shareOperationId> ?op } }`,
+    );
+    return result.type === 'bindings' ? result.bindings.map((row) => String(row['op'])).sort() : [];
+  }
+
+  async function opSubjectExists(store: OxigraphStore, subject: string): Promise<boolean> {
+    const result = await store.query(`ASK { GRAPH <${WS_META}> { <${subject}> ?p ?o } }`);
+    return result.type === 'boolean' && result.value;
+  }
+
+  it('preserves the local identity when the curator offers an equivalent operation id', async () => {
+    const store = new OxigraphStore();
+    stores.push(store);
+    await seedLocal(store);
+    // Content identical, so recovery classifies the KA as already
+    // materialized and SKIPS the graph replace. Pre-fix the bulk meta
+    // replacement then rotated the head to storage-ack-x and deleted the
+    // op-local operation subject — the exact identity a queued VM-publish
+    // job would have frozen at admission.
+    const result = await recoverContextGraphSwm(identityDeps(store, [...curatorEquivalent.meta]));
+    expect(result.completed).toBe(true);
+    expect(await headIds(store)).toEqual(['"op-local"']);
+    expect(await opSubjectExists(store, localShare.operationSubject)).toBe(true);
+    // The curator's operation subject lands as immutable history (same
+    // disposal as the public lane) — only its head-id row is withheld.
+    expect(await opSubjectExists(store, curatorEquivalent.operationSubject)).toBe(true);
+  });
+
+  it('still adopts the curator identity for a genuinely changed share (discriminator polarity)', async () => {
+    const store = new OxigraphStore();
+    stores.push(store);
+    await seedLocal(store);
+    // Same version, different digest: the marker-only predicate still skips
+    // the graph (it is digest-blind — the pre-existing F2 weakness), but the
+    // preservation decision compares the OPERATION rows and must refuse:
+    // curator authority for genuine changes is today's behavior, and
+    // preserving here could mask a real policy or content change.
+    const result = await recoverContextGraphSwm(identityDeps(store, [...curatorChanged.meta]));
+    expect(result.completed).toBe(true);
+    expect(await headIds(store)).toEqual(['"storage-ack-y"']);
+    expect(await opSubjectExists(store, localShare.operationSubject)).toBe(false);
+  });
+
+  it('still installs the curator head when the local head is absent (G7 repair preserved)', async () => {
+    const store = new OxigraphStore();
+    stores.push(store);
+    // Content present, head absent — the invisibility class #2050 G7 exists
+    // to repair. The preserve decision must NOT trigger (no healthy head), so
+    // the curator's meta is installed and the KA becomes readable again.
+    await store.insert(localShare.payload.map((quad) => ({ ...quad, graph: localShare.assertionGraph })));
+    const result = await recoverContextGraphSwm(identityDeps(store, [...curatorEquivalent.meta]));
+    expect(result.completed).toBe(true);
+    expect(await headIds(store)).toEqual(['"storage-ack-x"']);
+  });
+
+  it('replaces a multi-valued local head instead of preserving it (healthy-head conjunct)', async () => {
+    const store = new OxigraphStore();
+    stores.push(store);
+    await seedLocal(store);
+    await store.insert([{
+      subject: localShare.headSubject,
+      predicate: `${DKG}shareOperationId`,
+      object: '"phantom-op"',
+      graph: WS_META,
+    }]);
+    // A corrupt (multi-valued) head must not be preserved — the curator's
+    // authoritative meta replaces it, converging the head to one identity.
+    const result = await recoverContextGraphSwm(identityDeps(store, [...curatorEquivalent.meta]));
+    expect(result.completed).toBe(true);
+    expect(await headIds(store)).toEqual(['"storage-ack-x"']);
+  });
+
+  it('does not stack two equivalent curator head ids into the store (canonicalization)', async () => {
+    const store = new OxigraphStore();
+    stores.push(store);
+    // The parser ACCEPTS a payload whose head carries two byte-equivalent
+    // operation ids (storage-ACK + originator residue) and selects one; the
+    // raw insert must land only the SELECTED id or the freshly recovered
+    // head is born multi-valued.
+    const curatorTwin = swmFx.share({ version: 1, operationId: 'storage-ack-z', marker: 'identity', ual: UAL3 });
+    const payload = [
+      ...curatorEquivalent.meta,
+      ...curatorTwin.meta.filter((quad) => quad.subject === curatorTwin.operationSubject),
+      ...curatorTwin.meta.filter((quad) =>
+        quad.subject === curatorTwin.headSubject && quad.predicate === `${DKG}shareOperationId`),
+    ];
+    const result = await recoverContextGraphSwm(identityDeps(store, payload));
+    expect(result.completed).toBe(true);
+    expect((await headIds(store)).length).toBe(1);
+  });
+});
