+++ packages/agent/vitest.unit.config.ts
@@ -114,6 +114,7 @@ export default defineConfig({
       "test/rootless-durable-skips-legacy-partition.test.ts",
       "test/rootless-lifecycle-graph.test.ts",
       "test/swm-recovery.test.ts",
+      "test/swm-recovery-identity-preservation.test.ts",
       "test/dkg-agent-snapshot-store-injection.test.ts",
       "test/swm-snapshot-sync.test.ts",
       "test/sync-responder-protection.test.ts",
